requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions

absl-py==0.13.0
aiohttp==3.8.1
aiosignal==1.2.0
alpaca-trade-api==2.0.0
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.12
colorama==0.4.4
cycler==0.11.0
DateTime==4.4
deprecation==2.1.0
distlib==0.3.2
docopt==0.6.2
filelock==3.0.12
flatbuffers==1.12
fonttools==4.28.3
frozenlist==1.3.0
gast==0.4.0
google-auth==1.31.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==2.10
igraph==0.9.8
jsonpatch==1.32
jsonpointer==2.2
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.4
matplotlib==3.5.0
msgpack==1.0.3
multidict==6.0.2
neuralnet==0.2.0
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.2.4
Pillow==8.4.0
pipreqs==0.4.11
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
# pycairo==1.20.1
pyparsing==3.0.6
python-dateutil==2.8.1
python-igraph==0.9.8
pytz==2021.1
PyYAML==6.0
pyzmq==22.3.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.3
setuptools-scm==6.3.2
six==1.15.0
sortedcontainers==2.4.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
texttable==1.6.4
Theano==1.0.5
tomli==1.2.2
torchfile==0.1.0
tornado==6.1
tqdm==4.62.3
typing-extensions==3.7.4.3
urllib3==1.26.5
virtualenv==20.4.7
visdom==0.1.8.9
websocket-client==1.2.1
websockets==10.2
Werkzeug==2.0.1
wrapt==1.12.1
yarg==0.1.9
yarl==1.7.2
zope.interface==5.4.0
